requirements.txt
faster-whisper
torch
torchaudio
silero-vad
sounddevice
numpy
anthropic
httpx
pystray
Pillow
keyboard
python-dotenv
